Make `DartEnvironment` configurable from `DartExtension`
