+++ client/build.gradle.kts
@@ -63,18 +63,10 @@ dependencies {
 
 dart {
 
-//    publication {
-//        publicationDirectory = "..."
-//        pubExecutable = "..."
-//        pubSpec = "..."
-//    }
-
-
-//    environment {
-//        publicationDirectory = "..."
-//        pubExecutable = "..."
-//        pubSpec = "..."
-//    }
+    environment {
+        publicationDirectory = "${project.buildDir}/customPubPublication"
+        pubExecutable = "/usr/local/opt/dart/libexec/bin/${defaultEnvironment.pubExecutable}"
+    }
 
     tasks {
 
@@ -87,6 +79,7 @@ dart {
 //        }
 
     }
+
 }
 
 tasks.assemble {
